# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "财务资料与整改工作台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/gaps": {
      "post": {
        "summary": "New Gap",
        "operationId": "new_gap_api_v1_projects__project_id__gaps_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GapInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/issues": {
      "post": {
        "summary": "New Issue",
        "operationId": "new_issue_api_v1_projects__project_id__issues_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Issues",
        "operationId": "issues_api_v1_projects__project_id__issues_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "org_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Org Id"
            }
          },
          {
            "name": "overdue",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Overdue"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/issues/{issue_id}": {
      "get": {
        "summary": "Issue Detail",
        "operationId": "issue_detail_api_v1_issues__issue_id__get",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/issues/{issue_id}/extension-requests": {
      "post": {
        "summary": "Extension",
        "operationId": "extension_api_v1_issues__issue_id__extension_requests_post",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Issue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Extension"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/issues/{issue_id}/{action}": {
      "post": {
        "summary": "Issue Action",
        "operationId": "issue_action_api_v1_issues__issue_id___action__post",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Issue Id"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueActionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Extension": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "new_due": {
            "type": "string",
            "format": "date",
            "title": "New Due"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason",
          "expected_version",
          "new_due"
        ],
        "title": "Extension"
      },
      "GapInput": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "facts": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Facts"
          },
          "kind": {
            "type": "string",
            "enum": [
              "missing",
              "wrong_version",
              "scope_mismatch",
              "approval_missing",
              "control",
              "restricted",
              "other"
            ],
            "title": "Kind"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_id",
          "facts",
          "kind"
        ],
        "title": "GapInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IssueActionInput": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "evidence_version_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Version Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason",
          "expected_version"
        ],
        "title": "IssueActionInput"
      },
      "IssueCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 3,
            "title": "Title"
          },
          "facts": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Facts"
          },
          "kind": {
            "type": "string",
            "enum": [
              "missing",
              "wrong_version",
              "scope_mismatch",
              "approval_missing",
              "control",
              "restricted",
              "other"
            ],
            "title": "Kind"
          },
          "severity": {
            "type": "string",
            "enum": [
              "P0",
              "P1",
              "P2"
            ],
            "title": "Severity"
          },
          "org_id": {
            "type": "string",
            "title": "Org Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "verifier_id": {
            "type": "string",
            "title": "Verifier Id"
          },
          "due": {
            "type": "string",
            "format": "date",
            "title": "Due"
          },
          "gap_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Gap Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "facts",
          "kind",
          "severity",
          "org_id",
          "owner_id",
          "verifier_id",
          "due"
        ],
        "title": "IssueCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}